指令格式匹配正则表达式

+++ src/main/java/scau/os/soos/module/terminal/TerminalService.java
@@ -80,6 +80,7 @@ public class TerminalService {
 
 
     private String changeDirectory(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         try {
             if(FileController.getInstance().isExistedDirectory(arg)){
                 currentDirectory = arg;
@@ -91,6 +92,7 @@ public class TerminalService {
     }
 
     private String createFile(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         String path = currentDirectory + "/" + arg;
         try {
             FileController.getInstance().createFile(path);
@@ -105,6 +107,7 @@ public class TerminalService {
     }
 
     private String deleteFile(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         try {
             String path = currentDirectory + "/" + arg;
             FileController.getInstance().deleteFile(path);
@@ -121,6 +124,7 @@ public class TerminalService {
     }
 
     private String typeFile(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         try {
             String path = currentDirectory + "/" + arg;
             return FileController.getInstance().typeFile(path);
@@ -132,6 +136,7 @@ public class TerminalService {
     }
 
     private String copyFile(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+\\s+[^\\s]+$"))return "指令格式错误";
         // 分离源文件和目标文件
         String[] parts = arg.split(" ", 2);
         String src = currentDirectory + "/" + parts[0];
@@ -155,6 +160,7 @@ public class TerminalService {
     }
 
     private String makeDirectory(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         try {
             String path = currentDirectory + "/" + arg;
             FileController.getInstance().createDirectory(path);
@@ -169,6 +175,7 @@ public class TerminalService {
     }
 
     private String removeDirectory(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         try {
             String path = currentDirectory + "/" + arg;
             FileController.getInstance().deleteEmptyDirectory(path);
@@ -188,6 +195,7 @@ public class TerminalService {
 
 
     private String deleteDirectory(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         try {
             String path = currentDirectory + "/" + arg;
             FileController.getInstance().deleteFile(path);
@@ -204,6 +212,7 @@ public class TerminalService {
     }
 
     private String moveFile(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+\\s+[^\\s]+$"))return "指令格式错误";
         String[] parts = arg.split(" ", 2);
         String src = currentDirectory + "/" + parts[0];
         String dest = parts[1];
@@ -226,6 +235,7 @@ public class TerminalService {
     }
 
     private String changeFileAttribute(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+\\s+[^\\s]+\\s+[^\\s]+\\s+[^\\s]+\\s+[^\\s]+$"))return "指令格式错误";
         // 改变文件属性
         String[] parts = arg.split(" ", 5);
         String path = currentDirectory + "/" + parts[0];
@@ -233,7 +243,7 @@ public class TerminalService {
         for(int i = 1; i <=4; ++i)
         {
             if(!parts[i].matches("^(true|false)$"))
-                return "输入格式错误";
+                return "指令格式错误";
         }
         // 判断输入是否合法
         readOnly = Boolean.parseBoolean(parts[1]);
@@ -254,6 +264,7 @@ public class TerminalService {
     }
 
     private String formatDisk(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+$"))return "指令格式错误";
         // 磁盘格式化
         String[] parts = arg.split(" ", 2);
         String path = currentDirectory + "/" + parts[0];
@@ -268,6 +279,7 @@ public class TerminalService {
     }
 
     private String partitionDisk(String arg) {
+        if(!arg.matches("^\\s+[^\\s]+\\s+[^\\s]+\\s+\\d+$"))return "指令格式错误";
         // 磁盘分区
         String[] parts = arg.split(" ", 3);
         String src = parts[0];
@@ -276,7 +288,7 @@ public class TerminalService {
         try {
             size = Integer.parseInt(parts[2]);
         } catch (NumberFormatException e) {
-            return "输入格式错误";
+            return "指令格式错误";
         }
 
         try{
